requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
idna==2.9
multidict==4.7.5
Pillow==7.0.0
pycparser==2.20
python-chess==0.30.1
slackclient==2.5.0
tinycss2==1.0.2
webencodings==0.5.1
yarl==1.4.2
